CloseConfirmationDialog: pack custom content in content_area
Not message_area. That way, the scrolled window is expandable and fills
the available space when the dialog is resized.

But we need to set ourselves the spacing and margins, which is less
convenient and needs to be updated when GtkMessageDialog changes (or it
is possible to compute the total margins of message_area plus its
parents and apply the same, but to be future-proof it needs to take into
account all the ways to add extra space around a widget: margins,
padding, borders, the spacing between children in containers, etc, which
can be a bit tricky).